+++ src/com/intellij/idea/plugin/hybris/actions/AbstractExecuteAction.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
- * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -36,7 +36,7 @@ abstract class AbstractExecuteAction(
 
     override fun getActionUpdateThread() = ActionUpdateThread.BGT
 
-    protected open fun doExecute(consoleService: HybrisConsoleService) {
+    protected open fun doExecute(e: AnActionEvent, consoleService: HybrisConsoleService) {
         consoleService.executeStatement()
     }
 
@@ -65,7 +65,7 @@ abstract class AbstractExecuteAction(
         console.setInputText(content)
 
         invokeLater {
-            doExecute(consoleService)
+            doExecute(e, consoleService)
         }
     }