+++ src/com/intellij/idea/plugin/hybris/impex/actions/ImpExValidateAction.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
- * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -23,6 +23,7 @@ import com.intellij.idea.plugin.hybris.common.utils.HybrisIcons
 import com.intellij.idea.plugin.hybris.impex.file.ImpexFileType
 import com.intellij.idea.plugin.hybris.tools.remote.console.HybrisConsoleService
 import com.intellij.openapi.actionSystem.ActionUpdateThread
+import com.intellij.openapi.actionSystem.AnActionEvent
 
 class ImpExValidateAction : AbstractExecuteAction(
     ImpexFileType.defaultExtension,
@@ -38,7 +39,7 @@ class ImpExValidateAction : AbstractExecuteAction(
     }
 
     override fun getActionUpdateThread() = ActionUpdateThread.BGT
-    override fun doExecute(consoleService: HybrisConsoleService) {
+    override fun doExecute(e: AnActionEvent, consoleService: HybrisConsoleService) {
         consoleService.validateImpex()
     }
 }