+++ src/com/intellij/idea/plugin/hybris/settings/GroovySettings.kt
@@ -1,6 +1,6 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -22,4 +22,9 @@ data class GroovySettings(
     var enableActionsToolbar: Boolean = true,
     var enableActionsToolbarForGroovyTest: Boolean = false,
     var enableActionsToolbarForGroovyIdeConsole: Boolean = false,
+    var txMode: TransactionMode = TransactionMode.ROLLBACK,
 )
+
+enum class TransactionMode {
+    COMMIT, ROLLBACK
+}