+++ src/com/intellij/idea/plugin/hybris/settings/components/DeveloperSettingsComponent.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
- * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -29,7 +29,7 @@ import com.intellij.util.xmlb.XmlSerializerUtil
     storages = [Storage(value = STORAGE_HYBRIS_DEVELOPER_SPECIFIC_PROJECT_SETTINGS, roamingType = RoamingType.DISABLED)]
 )
 @Service(Service.Level.PROJECT)
-class DeveloperSettingsComponent(private val project: Project) : PersistentStateComponent<DeveloperSettings> {
+class DeveloperSettingsComponent : PersistentStateComponent<DeveloperSettings> {
 
     private val state = DeveloperSettings()