+++ src/com/intellij/idea/plugin/hybris/toolwindow/RemoteHacConnectionDialog.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
- * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -150,7 +150,7 @@ class RemoteHacConnectionDialog(
             }.layout(RowLayout.PARENT_GRID)
 
             row {
-                label("Session Cookie Name:")
+                label("Session Cookie name:")
                 sessionCookieNameTextField = textField()
                     .comment("Optional: override the session cookie name. Default is JSESSIONID.")
                     .align(AlignX.FILL)