+++ CHANGELOG.md
@@ -23,6 +23,9 @@
 - Added API to retrieve item type and expression of the bind parameter [#1420](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1420)
 - Improved code formatting for expressions within `WHERE` clause [#1421](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1421)
 
+### `Groovy` enhancements
+- Adjusted execution mode selection `commit`/`rollback` [#1422](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1422)
+
 ## [2025.1.4]
 
 ### `ImpEx` enhancements